Refactor hud

+++ src/com/github/nathankuhn/circuitgame/utils/Vector2i.java
@@ -1,19 +1,19 @@
 package com.github.nathankuhn.circuitgame.utils;
 
-public class Vector2f {
+public class Vector2i {
 
-    public float x;
-    public float y;
+    public int x;
+    public int y;
 
-    public Vector2f() {
-         x = 0;
-         y = 0;
+    public Vector2i() {
+        x = 0;
+        y = 0;
     }
-    public Vector2f(float x, float y) {
+    public Vector2i(int x, int y) {
         this.x = x;
         this.y = y;
     }
-    public Vector2f(Vector3f vector) {
+    public Vector2i(Vector3i vector) {
         this.x = vector.x;
         this.y = vector.y;
     }
